requirements.txt
tensorflow==2.18.0
numpy==1.26.4
pillow==10.3.0
flask==3.0.3
opencv-python==4.10.0.84
mtcnn
pycryptodome